requirements.txt
requests==2.26.0
environs==9.3.5
python-telegram-bot==13.8.1